autoboxing, casting
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@24473 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/client/async/SplitFileFetcherSubSegment.java
@@ -788,8 +788,8 @@ public class SplitFileFetcherSubSegment extends SendableGet implements SupportsB
 		boolean found = false;
 		synchronized(segment) {
 			for(int i=0;i<blockNums.size();i++) {
-				Integer token = (Integer) blockNums.get(i);
-				int num = ((Integer)token).intValue();
+				Integer token = blockNums.get(i);
+				int num = token;
 				if(num == blockNum) {
 					blockNums.remove(i);
 					if(logMINOR) Logger.minor(this, "Removed block "+blockNum+" from "+this);
